+++ cmd/config_set.go
@@ -15,21 +15,21 @@ var configSetCmd = &cobra.Command{
 	Long: `Updates the configuration. Set the flags for the options
 you want to change. Other options will remain unchanged.`,
 	Args: cobra.NoArgs,
-	RunE: python(func(cmd *cobra.Command, _ []string, d *pythonData) error {
+	RunE: withStore(func(cmd *cobra.Command, _ []string, st *store) error {
 		flags := cmd.Flags()
 
 		// Read existing config
-		set, err := d.store.Settings.Get()
+		set, err := st.Settings.Get()
 		if err != nil {
 			return err
 		}
 
-		ser, err := d.store.Settings.GetServer()
+		ser, err := st.Settings.GetServer()
 		if err != nil {
 			return err
 		}
 
-		auther, err := d.store.Auth.Get(set.AuthMethod)
+		auther, err := st.Auth.Get(set.AuthMethod)
 		if err != nil {
 			return err
 		}
@@ -41,21 +41,21 @@ you want to change. Other options will remain unchanged.`,
 		}
 
 		// Save updated config
-		err = d.store.Auth.Save(auther)
+		err = st.Auth.Save(auther)
 		if err != nil {
 			return err
 		}
 
-		err = d.store.Settings.Save(set)
+		err = st.Settings.Save(set)
 		if err != nil {
 			return err
 		}
 
-		err = d.store.Settings.SaveServer(ser)
+		err = st.Settings.SaveServer(ser)
 		if err != nil {
 			return err
 		}
 
 		return printSettings(ser, set, auther)
-	}, pythonConfig{}),
+	}, storeOptions{}),
 }